tests: moved utils tests to mod
